Add the answer model, contoller and view to each question

+++ app/Http/Controllers/QuestionsController.php
@@ -8,6 +8,9 @@ use App\Http\Requests\AskQuestionRequest;
 
 class QuestionsController extends Controller
 {
+    public function __construct(){
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -38,6 +41,7 @@ class QuestionsController extends Controller
      */
     public function store(AskQuestionRequest $request)
     {
+        //$request->user()->questions()->create($request->only('title', 'body'));
         $request->user()->questions()->create($request->only('title', 'body'));
 
         return redirect()->route('questions.index')->with('success', 'Your data was submited');
@@ -64,7 +68,11 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        return view('questions.edit', compact('question'));
+        // if(\Gate::denies('update-question', $question)){
+        //     abort(403, 'Access denied');
+        // }
+        $this->authorize('update', $question);
+         return view('questions.edit', compact('question'));
     }
 
     /**
@@ -76,6 +84,11 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        // if(\Gate::denies('update-question', $question)){
+        //     abort(403, 'Access denied');
+        // }
+        $this->authorize('update', $question);
+
        $question->update($request->only('title', 'body'));
 
         return redirect()->route('questions.index')->with('success', 'Your question was updated');
@@ -91,6 +104,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        // if(\Gate::denies('delete-question', $question)){
+        //     abort(403, 'Access denied');
+        // }
+        $this->authorize('delete', $question);
         $question->delete();
 
         return redirect()->route('questions.index')->with('success', 'Your question was deleted');
